refactor: device관련 hook 파일명 수정
- useIsDesktop -> useDevice
