+++ src/components/Footer/Footer.tsx
@@ -1,7 +1,7 @@
 import { useRouter } from 'next/router';
 import { ReactComponent as ArrowRight } from '@src/assets/icons/arrow-right-16x16.svg';
 import Channels from '@src/components/Footer/Channels';
-import { useIsDesktop } from '@src/hooks/useIsDesktop';
+import { useIsDesktop } from '@src/hooks/useDevice';
 import * as S from './Footer.style';
 
 export function Footer() {
