+++ src/components/Header/Header.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { useIsDesktop } from '@src/hooks/useIsDesktop';
+import { useIsDesktop } from '@src/hooks/useDevice';
 import DesktopHeader from './Desktop/DesktopHeader';
 import MobileHeader from './Mobile/MobileHeader';
