+++ src/hooks/useDevice.ts
